아이템 기능 개발완료>>
아이템 기능 개발완료>>

+++ Assets/Scripts/Player/Player.cs
@@ -357,7 +357,7 @@ public class Player : PlayerComponent
 
         IsAccelerateThisFrame = isAccelerate;
     }
-    public void UpdateHealth(int health_)
+    public void UpdateHealth(float health_)
     {
         if(health_ > 0)
         {
